Minor API usability improvements.
